Organize Theme Files (#304)
