+++ services/ui-src/src/styles/components/divider.ts
@@ -1,0 +1,11 @@
+import { ComponentStyleConfig } from "@chakra-ui/react";
+
+const baseStyles = {
+  borderColor: "palette.gray_lighter",
+};
+
+const dividerTheme: ComponentStyleConfig = {
+  baseStyle: baseStyles,
+};
+
+export default dividerTheme;
