+++ services/ui-src/src/styles/components/index.ts
@@ -1,0 +1,21 @@
+import accordionTheme from "./accordion";
+import buttonTheme from "./button";
+import dividerTheme from "./divider";
+import headingTheme from "./heading";
+import linkTheme from "./link";
+import listTheme from "./list";
+import modalTheme from "./modal";
+import tableTheme from "./table";
+import textTheme from "./text";
+
+export const components = {
+  Accordion: accordionTheme,
+  Button: buttonTheme,
+  Divider: dividerTheme,
+  Heading: headingTheme,
+  Link: linkTheme,
+  List: listTheme,
+  Modal: modalTheme,
+  Table: tableTheme,
+  Text: textTheme,
+};
