+++ services/ui-src/src/styles/foundations/index.ts
@@ -1,0 +1,13 @@
+import { breakpoints } from "./breakpoints";
+import { fonts } from "./fonts";
+import { lineHeights } from "./lineHeights";
+import { colors } from "./colors";
+import { sizes } from "./sizes";
+
+export const foundations = {
+  breakpoints: breakpoints,
+  colors: colors,
+  fonts: fonts,
+  lineHeights: lineHeights,
+  sizes: sizes,
+};
