+++ services/ui-src/src/styles/styles.ts
@@ -1,0 +1,7 @@
+export const styles = {
+  global: {
+    body: {
+      color: "palette.base",
+    },
+  },
+};
